Docker on Python 3.12, Alembic, calibration UI, AI summary, LLM tests
Docker was the highest-risk unverified claim and it was broken three ways.

1. docker-compose.yml did not parse at all: `_comment` is not a valid root
   property, so the README's first command failed immediately.
2. The API image used the Playwright base, which ships Python 3.10. The app
   needs 3.12 (StrEnum). It built cleanly and died at import. The cause was
   that the Dockerfile hand-duplicated the dependency list instead of
   installing from pyproject, so `requires-python = ">=3.12"` was never
   enforced. It now installs from pyproject and asserts the interpreter at
   build time. The image also dropped from 3.75 GB to 662 MB.
3. alembic.ini and migrations/ were not copied into the image.

Alembic is now configured with an initial autogenerated migration, and
tests/unit/test_schema_parity.py exists - the guard db/session.py has been
claiming for several commits. It builds one database from migrations and one
from ORM metadata and compares tables, columns, types, nullability, primary
keys, indexes and foreign keys.

/dev/roof-calibration implements the full tool: source-pixel editing, add and
drag and rename vertices, draw edges with type assignment, build facets,
assign an eave per facet, layer toggles, live measurements, validation,
import/export/copy JSON and reset to committed. Every click converts to
source-map pixels before it reaches state, so pan and zoom cannot leak into
the data - the exact bug that once offset the whole calibration.

ai_summary is now generated rather than merely plumbed. The model may write
prose but not produce a number: every figure in the returned text must match
a value the backend computed, or the whole summary is discarded for the
deterministic template. Discarding all of it is deliberate - if one figure
was fabricated there is no way to tell which sentences were too.

Fixes a regression this introduced: finalize generated the summary before
checking readiness, turning a clean 409 into a 500.

430 API tests, MyPy strict clean across 39 files, Ruff clean.

+++ apps/api/app/core/config.py
@@ -142,7 +142,7 @@ def meters_per_source_pixel(latitude_deg: float, zoom: int, scale: int) -> float
         * circumference
         / (WEB_MERCATOR_TILE_SIZE_PX * (2**zoom))
     )
-    return logical / scale
+    return float(logical / scale)
 
 
 class Settings(BaseSettings):
